pyproject.toml
[project]
name = "multi-translate-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.3",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.28",
    "openai-whisper>=20250625",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
]
